+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
-    verbose: true,
+    verbose: false,
     moduleFileExtensions: ['ts', 'js', 'json', 'node'],
     testMatch: ['**/__tests__/**/*.test.ts']
 };